requirements.txt
fairbench==0.7.4
gdown==5.2.0
matplotlib==3.8.1
numpy==1.26.4
ollama==0.3.3
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
scikit_learn==1.5.1
scipy==1.13.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.48.2
wandb==0.18.2
yacs==0.1.8
git+https://github.com/xinyu1205/recognize-anything.git
tensorboard
fairscale
datasets==3.6.0
sentence_transformers==4.1.0
timm==1.0.15
pyts==0.13.0
lightly==1.5.20